update(tt): organizing the components and hooks
