+++ apps/total-typescript/src/hooks/use-survey.tsx
@@ -41,7 +41,7 @@ export const SurveyProvider: React.FC<
   const [currentQuestion, setCurrentQuestion] = React.useState<string>('')
 
   const question = useQuestion({
-    currentQuestion: survey.questions[currentQuestion],
+    currentOffer: survey.questions[currentQuestion],
     currentQuestionKey: currentQuestion,
     questionSet: survey.questions,
     config: surveyConfig,
