+++ packages/quiz/src/hooks/use-question.test.tsx
@@ -36,7 +36,7 @@ describe('can render questions', () => {
     const {
       result: {current},
     } = renderHook(() =>
-      useQuestion({currentQuestion: ESSAY_QUESTION, questionSet: {}}),
+      useQuestion({currentOffer: ESSAY_QUESTION, questionSet: {}}),
     )
 
     expect(current).not.toBeUndefined()
@@ -51,7 +51,7 @@ describe('can render questions', () => {
     const {
       result: {current},
     } = renderHook(() =>
-      useQuestion({currentQuestion: MULTIPLE_CHOICE_QUESTION, questionSet: {}}),
+      useQuestion({currentOffer: MULTIPLE_CHOICE_QUESTION, questionSet: {}}),
     )
     expect(current).not.toBeUndefined()
     current && render(<Essay question={current} />)
